Configurazione di connessione al SQL Server.

+++ ArticoliWebService/Models/Articoli.cs
@@ -22,7 +22,7 @@ namespace ArticoliWebService.Models
         public string? CodStat { get; set; }
 
         [Range(0, 100, ErrorMessage = "I pezzi per cartone devono essere compresi tra 0 e 100.")]
-        public int? PzCart { get; set; }
+        public Int16? PzCart { get; set; }
 
         [Range(0.01, 100, ErrorMessage = "Il peso netto deve essere compreso tra 0.01 e 100.")]
         public double? PesoNetto { get; set; }
